fix: Random failure only on Amazon
Tried to reproduce the issue locally but it never popped up. When debugging on Amazon it'd randomly happen. `interaction.reply` seemed to not be set when a request lingers for a bit, so it's possible it's some sort of network latency issue.
I've added some more error handling, time will tell if it'll be enough.